Read event offset from google calendar api
Previously the offset was generated from a string in the event title,
this allowed an event to fire at a time before or after the actual
event start. This commit uses the reminder settings for the event
to determine when the offset is reached. If no reminders are set
then the default for the calendar is used. If multiple reminders
are set, then then largest offset is used.